scripts/collection/informal_scripts/Twitter Data Scraping Script/pyproject.toml
[tool.poetry]
name = "twitter_script"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
searchtweets-v2 = "^1.1.1"
numpy = "^1.22.4"
pandas = "^1.4.2"
requests = "^2.27.1"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
